mx-actor-manager: unparent container before removing children
When removing a container from the stage, if we remove all its
children before removing the container itself we end up triggering a
relayout of the tree for each removal.

This patch propose an improvment by unparenting the container first,
before removing all its children.

Also fix a bug with the callback not being recalled when there are
still operations to process.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2662