tooltip: Update to support paint-volumes
Tooltips are a bit tricky in that we always want them aligned with stage
coordinates instead of being transform according the actor they apply
too, we also want tooltips to be painted on top of everything else.

Tooltips subclass MxFloatingWidget to ensure they are painted on top of
everything and that works by painting after all signal handlers of the
stage's "paint" signal instead of as part of the standard paint-cycle.

The thing that was making paint-volumes tricky though was that tooltips
were parented onto the MxWidget to which the tip applies, even though we
want them to paint in stage coordinates. It turns out to be rather
tricky to implement the apply_transform vfunc this way, since it
involves calculating the inverse transform of the parent's
stage-relative transform so we can negate it and get back to stage
coordinates. The other tricky thing is that when actually painting
the tooltip then we are actually in stage coordinates and so we don't
have to negate the parent transforms in that case.

This patch makes is so that tooltips are parented on the stage
so now implementing apply_transform is trivial. The trade off with this
is a bit more complexity to handle the tooltip possibly being destroyed
out-of-sync with the widget being destroyed and that we don't
necessarily have a stage while a widget is being constructed.

Tooltips are now only instantiated lazily during mx_widget_show_tooltip
if there is a corresponding stage. Since we now need to be able to set
tooltip text before creating a tooltip widget, the text is now saved
as part of the widget.

This patch also bumps MX's clutter version dependency to 1.6.17 since
we depend on ClutterStage being able to safely handle children being
removed while destroying children.