image: give correct rowstride argument to cogl_texture_set_region
On the Sodaville platform, we experienced a memory corruption in media
explorer triggered by Cogl when premultiplying a texture. It turns out
that we gave a wrong rowstride (8 bytes smaller) to Cogl when calling
cogl_texture_set_region. Because Cogl relies on rowstride to allocate
its buffers, but uses width * pixelsize for processing images, we were
writing 8 bytes behind an allocated buffer and corrupting malloc's
internal datas.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@linux.intel.com>