stack: add 'crop' child property
The 'crop' child property attempts to allocate the actor in such way
that it will completely fill the available space while keeping the
aspect ratio of the child, even if some area of the actor are over the
allocation of its parent. In such case the paint uses a clip rectangle
to not draw over the parent's allocation.