dialog: Use the portable Cogl wrappers for the GLSL builtins
The blur shader used in MxDialog was trying to use different GLSL
builtin names depending on whether Mx is being built for GLES2 or
GL. However it was detecting this with COGL_HAS_GLES2. Now that Cogl
supports multiple drivers in one binary, COGL_HAS_GLES2 doesn't
necessarily imply that Cogl is actually running on GLES2. Cogl now has
portable wrappers for most of the GLSL builtins so there should no
longer be any need to use different code depending on the GL
driver. This patch changes it to use those names.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2684