viewport: implement apply_transform and get_paint_volume
Use the apply_transform function to translate the actor, rather than
translating manually in the paint function, and implement
get_paint_volume so that the paint volume correctly matches the area
that will be painted. The "clip-to-allocation" property also no longer
needs to be set since MxScrollView will now clip the child itself. This
fixes the issue where clip-to-allocation was clipping the wrong region
since it did not take the scrolling translation into account.