notebook: rework API and update to Clutter 1.10 API
The add_actor/remove_actor/foreach vfunc from ClutterContainer have
been deprecated and should not be used anymore.

Also the new clutter API make it harder to support the current API
using ClutterActor* as index in the list of children/pages. This patch
rework the API to be closer to GtkNotebook using integer to design
children/pages.