dialog: remove use of deprecated Clutter API
Remove the use of ClutterShader, which has been deprecated in Clutter.
This means that the dialog no longer has a blurred background and that
MxOffscreen and the associated hooks into the mapped, allocate, paint and
pick signals are no longer required.

Replace the use of deprecated ClutterContainer API with equivalent
ClutterActor API.