Remove the use of ClutterTexture
ClutterTexture has been deprecated, so this patch removes API that exposes
ClutterTexture and replaces the use of ClutterTexture internally. In most
cases, ClutterTexture can be replaced by CoglTexture, which should also
improve performance.

This patch also removes the MxTextureFrame actor (which depended on
ClutterTexture) since it was not useful on its own and can be
implemented in other actors by using the mx_texture_frame_paint_texture
utility function.