kinetic-scroll-view: Replace motion buffer with a moving average
The motion buffer stored each motion event which happened in a sequence
(for example, a single swipe) on the kinetic scroll view, in order that
the average event coordinates and time could be computed at the end of
the sequence (on releasing the touchscreen). This could lead to problems
with very long event sequences — potentially allocating a huge event
array and leading to allocation failure.

Avoid that by computing the average of the events as they are received.

https://github.com/clutter-project/mx/pull/98